+++ include/ERG_Mode.h
@@ -25,10 +25,11 @@ class ErgMode {
   // What used to be in the ERGTaskLoop(). This is the main control function for ERG Mode and the powertable operations.
   void runERG();
   void computeErg();
+  void onFTMSCommand(uint8_t opcode);
   void _writeLog(float currentIncline, float newIncline, int currentSetPoint, int newSetPoint, int currentWatts, int newWatts, int currentCadence, int newCadence);
 
  private:
-  bool engineStopped = false;
+  bool resumeErgOnCadence = false;
 
   int mode = Mode::MAINTAIN;
   Measurement prevWatts;